requirements.txt
django-admin==2.0.2
Django==4.2.4
djangorestframework==3.14.0
selenium==4.11.2
colorlog==6.7.0